Require C11
Error out if the compiler does not support C11.

Removes support for `__thread` since we now rely on C11 `_Thread_local`.

*Auxilliary fixes required to build with C11:*

Replace typeof() with size_t in loop header
All uses of the `PREPARE_REQUESTS_WITH_NO_FREE` macro map to size_t.

Add feature test macro _POSIX_C_SOURCE to enable fileno()
Flex prior to 2.6.6 uses fileno() but does not define the needed
feature test macros.
See https://github.com/westes/flex/issues/263 for details.

Add feature test macro to enable pthread_mutexattr_settype()

Enable nanosleep with _POSIX_C_SOURCE
Move calls to nanosleep out of headers and add the feature test macro
_POSIX_C_SOURCE to enable its use. This should not cause
any significant overheads.

Reduce_local test: Add feature test macros
Needed for strsep(), posix_memalign(), and getopt().
Also fix the min/max macros. There is no need to cache the variables
as these macros are used without producing side-effects.

Include ompi_config.h in reduce_local.c
