+++ opal/class/opal_lifo.c
@@ -19,7 +19,11 @@
  * $HEADER$
  */
 
+/* needed for nanosleep() */
+#define _POSIX_C_SOURCE 200809L
+
 #include "opal_config.h"
+#include <time.h>
 #include "opal/class/opal_lifo.h"
 
 static void opal_lifo_construct(opal_lifo_t *lifo)
@@ -31,3 +35,16 @@ static void opal_lifo_construct(opal_lifo_t *lifo)
 }
 
 OBJ_CLASS_INSTANCE(opal_lifo_t, opal_object_t, opal_lifo_construct, NULL);
+
+
+void opal_lifo_release_cpu(void)
+{
+    /* NTH: there are many ways to cause the current thread to be suspended. This one
+     * should work well in most cases. Another approach would be to use poll (NULL, 0, ) but
+     * the interval will be forced to be in ms (instead of ns or us). Note that there
+     * is a performance improvement for the lifo test when this call is made on detection
+     * of contention but it may not translate into actually MPI or application performance
+     * improvements. */
+    static struct timespec interval = {.tv_sec = 0, .tv_nsec = 100};
+    nanosleep(&interval, NULL);
+}
